+++ src/Whim.TreeLayout.Tests/PopulatedTreeLayoutEngine.cs
@@ -1,4 +1,5 @@
 using Moq;
+using Xunit;
 
 namespace Whim.TreeLayout.Tests;
 
@@ -12,31 +13,31 @@ internal class TestTreeEngine
 	public Mock<IConfigContext> ConfigContext = new();
 
 	public Mock<IWindow> LeftWindow = new();
-	public LeafNode? LeftNode;
+	public LeafNode LeftNode;
 
 	public Mock<IWindow> RightTopLeftTopWindow = new();
-	public LeafNode? RightTopLeftTopNode;
+	public LeafNode RightTopLeftTopNode;
 
 	public Mock<IWindow> RightBottomWindow = new();
-	public LeafNode? RightBottomNode;
+	public LeafNode RightBottomNode;
 
 	public Mock<IWindow> RightTopRight1Window = new();
-	public LeafNode? RightTopRight1Node;
+	public LeafNode RightTopRight1Node;
 
 	public Mock<IWindow> RightTopRight2Window = new();
-	public LeafNode? RightTopRight2Node;
+	public LeafNode RightTopRight2Node;
 
 	public Mock<IWindow> RightTopRight3Window = new();
-	public LeafNode? RightTopRight3Node;
+	public LeafNode RightTopRight3Node;
 
 	public Mock<IWindow> RightTopLeftBottomLeftWindow = new();
-	public LeafNode? RightTopLeftBottomLeftNode;
+	public LeafNode RightTopLeftBottomLeftNode;
 
 	public Mock<IWindow> RightTopLeftBottomRightTopWindow = new();
-	public LeafNode? RightTopLeftBottomRightTopNode;
+	public LeafNode RightTopLeftBottomRightTopNode;
 
 	public Mock<IWindow> RightTopLeftBottomRightBottomWindow = new();
-	public LeafNode? RightTopLeftBottomRightBottomNode;
+	public LeafNode RightTopLeftBottomRightBottomNode;
 	public TreeLayoutEngine Engine;
 
 	public TestTreeEngine()
@@ -56,38 +57,47 @@ internal class TestTreeEngine
 		RightTopLeftBottomRightTopWindow.Setup(m => m.ToString()).Returns("rightTopLeftBottomRightTopWindow");
 		RightTopLeftBottomRightBottomWindow.Setup(m => m.ToString()).Returns("rightTopLeftBottomRightBottomWindow");
 
-		LeftNode = Engine.AddWindow(LeftWindow.Object);
+		LeftNode = Engine.AddWindow(LeftWindow.Object)!;
+		Assert.NotNull(LeftNode);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(LeftWindow.Object);
 
-		RightTopLeftTopNode = Engine.AddWindow(RightTopLeftTopWindow.Object);
+		RightTopLeftTopNode = Engine.AddWindow(RightTopLeftTopWindow.Object)!;
+		Assert.NotNull(RightTopLeftTopNode);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopLeftTopWindow.Object);
 
 		Engine.Direction = NodeDirection.Down;
-		RightBottomNode = Engine.AddWindow(RightBottomWindow.Object);
+		RightBottomNode = Engine.AddWindow(RightBottomWindow.Object)!;
+		Assert.NotNull(RightBottomNode);
 
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopLeftTopWindow.Object);
 		Engine.Direction = NodeDirection.Right;
 
-		RightTopRight1Node = Engine.AddWindow(RightTopRight1Window.Object);
+		RightTopRight1Node = Engine.AddWindow(RightTopRight1Window.Object)!;
+		Assert.NotNull(RightTopRight1Node);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopRight1Window.Object);
 		Engine.Direction = NodeDirection.Down;
 
-		RightTopRight2Node = Engine.AddWindow(RightTopRight2Window.Object);
+		RightTopRight2Node = Engine.AddWindow(RightTopRight2Window.Object)!;
+		Assert.NotNull(RightTopRight2Node);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopRight2Window.Object);
 
-		RightTopRight3Node = Engine.AddWindow(RightTopRight3Window.Object);
+		RightTopRight3Node = Engine.AddWindow(RightTopRight3Window.Object)!;
+		Assert.NotNull(RightTopRight3Node);
 
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopLeftTopWindow.Object);
 		Engine.Direction = NodeDirection.Down;
 
-		RightTopLeftBottomLeftNode = Engine.AddWindow(RightTopLeftBottomLeftWindow.Object);
+		RightTopLeftBottomLeftNode = Engine.AddWindow(RightTopLeftBottomLeftWindow.Object)!;
+		Assert.NotNull(RightTopLeftBottomLeftNode);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopLeftBottomLeftWindow.Object);
 		Engine.Direction = NodeDirection.Right;
 
-		RightTopLeftBottomRightTopNode = Engine.AddWindow(RightTopLeftBottomRightTopWindow.Object);
+		RightTopLeftBottomRightTopNode = Engine.AddWindow(RightTopLeftBottomRightTopWindow.Object)!;
+		Assert.NotNull(RightTopLeftBottomRightTopNode);
 		ActiveWorkspace.Setup(x => x.FocusedWindow).Returns(RightTopLeftBottomRightTopWindow.Object);
 		Engine.Direction = NodeDirection.Down;
 
-		RightTopLeftBottomRightBottomNode = Engine.AddWindow(RightTopLeftBottomRightBottomWindow.Object);
+		RightTopLeftBottomRightBottomNode = Engine.AddWindow(RightTopLeftBottomRightBottomWindow.Object)!;
+		Assert.NotNull(RightTopLeftBottomRightBottomNode);
 	}
 }
